requirements.txt
# SRE Copilot - Dependencies
# ============================

# Core CrewAI framework
crewai>=0.28.0
crewai-tools>=0.1.0

# Language model support
langchain>=0.1.0
openai>=1.3.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

# Configuration and data handling
pyyaml>=6.0
python-dotenv>=1.0.0

# Kubernetes integration
kubernetes>=28.1.0

# Datadog integration
datadog-api-client>=2.20.0

# Communication tools
slack-sdk>=3.21.0
pymsteams>=0.2.0

# Utilities
requests>=2.31.0
click>=8.1.0
rich>=13.6.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
